+++ src/Zphil.ReSharperCli/Program.cs
@@ -36,6 +36,8 @@ builder.Services.AddSingleton<IEnvironment, SystemEnvironment>();
 builder.Services.AddSingleton<IProcessRunner, ProcessRunner>();
 builder.Services.AddSingleton<JbLocator>();
 builder.Services.AddSingleton<ConfigResolver>();
+builder.Services.AddSingleton(_ => new JbRunLock());
+builder.Services.AddSingleton<JbRunner>();
 builder.Services.AddSingleton<InspectService>();
 builder.Services.AddSingleton<CleanupService>();